# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Course Planner API
  version: 1.0.0
  description: |
    Shared HTTP contract for the Course Planner system.
    All absolute datetimes are RFC3339 UTC. User-facing local time is interpreted
    in the authenticated user's immutable IANA timezone.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /calendars/{calendarId}/periods:
    get:
      operationId: listPeriods
      parameters:
        - { $ref: '#/components/parameters/CalendarId' }
      responses:
        '200': { description: Period list }
    post:
      operationId: createPeriod
      parameters:
        - { $ref: '#/components/parameters/CalendarId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PeriodRequest' }
      responses:
        '200': { description: Created period }
  /calendars/{calendarId}/periods/{periodId}:
    patch:
      operationId: updatePeriod
      parameters:
        - { $ref: '#/components/parameters/CalendarId' }
        - { $ref: '#/components/parameters/PeriodId' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PeriodRequest' }
      responses:
        '200': { description: Updated period }
    delete:
      operationId: deletePeriod
      parameters:
        - { $ref: '#/components/parameters/CalendarId' }
        - { $ref: '#/components/parameters/PeriodId' }
      responses:
        '204': { description: Deleted }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CalendarId:
      name: calendarId
      in: path
      required: true
      schema: { type: string, format: uuid }
    PeriodId:
      name: periodId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    PeriodRequest:
      type: object
      required: [periodNo, startLocal, endLocal]
      properties:
        periodNo: { type: integer, minimum: 1 }
        startLocal: { type: string, pattern: '^([01]\\d|2[0-3]):[0-5]\\d$' }
        endLocal: { type: string, pattern: '^([01]\\d|2[0-3]):[0-5]\\d$' }
...